resolve comments and refactor code
