+++ serverless/executor-enclave/supervisord.conf
@@ -41,7 +41,7 @@ stderr_logfile_maxbytes=0
 
 # Oyster secret store
 [program:oyster-secret-store]
-command=/app/oyster-secret-store --inject-port 6002 --config-file /etc/oyster_secret_store_config.json
+command=/app/oyster-secret-store --external-port 6002 --config-file /etc/oyster_secret_store_config.json
 autorestart=true
 stdout_logfile=/dev/stdout
 stdout_logfile_maxbytes=0
