+++ serverless/executor/src/lib.rs
@@ -1,7 +1,9 @@
 pub mod cgroups;
 pub mod config;
+pub mod constant;
 pub mod event_handler;
 pub mod job_handler;
+pub mod model;
 pub mod node_handler;
 mod tests;
 pub mod timeout_handler;
