+++ serverless/secret-store/oyster_secret_store_config.json
@@ -3,7 +3,7 @@
     "secret_store_path": "/app/store/",
     "common_chain_id": 421614,
     "http_rpc_url": "https://sepolia-rollup.arbitrum.io/rpc",
-    "web_socket_url": "wss://arbitrum-sepolia.infura.io/ws/v3/",
+    "web_socket_url": "wss://arb-sepolia.g.alchemy.com/v2/",
     "tee_manager_contract_addr": "0xBce5fba7361A2E6Ca5BBe624cC76B6f6fe45b1C8",
     "secret_manager_contract_addr": "0x6cc663135635c71175a35E4710fC7Ef4e12a085b",
     "enclave_signer_file": "/app/ecdsa.sec",
